lambda-gcp-vision/requirements.txt
boto3
google-cloud-vision==3.4.4
PyMuPDF==1.24.1